renamer fil: Behandlingsårsak.ts -> behandlingsårsak.ts (liten forbokstav)
