+++ src/frontend/Komponenter/Behandling/Fanemeny/KorrigeringFane.tsx
@@ -4,7 +4,7 @@ import { useBehandling } from '../../../App/context/BehandlingContext';
 import { AlertInfo } from '../../../Felles/Visningskomponenter/Alerts';
 import { skalFerdigstilleUtenBeslutter } from '../VedtakOgBeregning/Felles/utils';
 import DataViewer from '../../../Felles/DataViewer/DataViewer';
-import { Behandlingsårsak } from '../../../App/typer/Behandlingsårsak';
+import { Behandlingsårsak } from '../../../App/typer/behandlingsårsak';
 import { Behandling } from '../../../App/typer/fagsak';
 
 interface Props {
